HV: include: integral type cleanup
This patch cleans up the integral type violations to MISRA C rules, mostly
related to signed constants that should be unsigned but also spelling out two
integer narrowing and dropping some macros negating unsigned integers.

v1 -> v2:

    * Drop INT_ROUNDUPx macros since they are never used.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/arch/x86/cpu.h
@@ -462,8 +462,8 @@ write_xcr(int reg, uint64_t val)
 {
 	uint32_t low, high;
 
-	low = val;
-	high = val >> 32;
+	low = (uint32_t)val;
+	high = (uint32_t)(val >> 32);
 	asm volatile("xsetbv" : : "c" (reg), "a" (low), "d" (high));
 }
 #else /* ASSEMBLER defined */
